requirements.txt
scrapy
beautifulsoup4
pymysql
psycopg2
pymongo
pandas
docker
